+++ PGS4NET/Segments/CompositionObject.cs
@@ -26,13 +26,13 @@ public class CompositionObject
     ///     The horizontal offset of the object's top-left corner within the screen. If the
     ///     object is cropped, then this applies only to the visible area.
     /// </summary>
-    public ushort X { get; set; }
+    public int X { get; set; }
 
     /// <summary>
     ///     The vertical offset of the object's top-left corner within the screen. If the
     ///     object is cropped, then this applies only to the visible area.
     /// </summary>
-    public ushort Y { get; set; }
+    public int Y { get; set; }
 
     /// <summary>
     ///     Whether or not the composition object is forced. This is typically used to translate
@@ -75,7 +75,7 @@ public class CompositionObject
     ///     If set, defines the visible area of the object. Otherwise, the entire object is
     ///     shown.
     /// </param>
-    public CompositionObject(CompositionId id, ushort x, ushort y, bool forced, Crop? crop)
+    public CompositionObject(CompositionId id, int x, int y, bool forced, Crop? crop)
     {
         Id = id;
         X = x;
