+++ PGS4NET/Segments/InitialObjectDefinitionSegment.cs
@@ -24,17 +24,17 @@ public class InitialObjectDefinitionSegment : ObjectDefinitionSegment
     /// <summary>
     ///     The declared length of this object's data buffer, including all follow-on portions.
     /// </summary>
-    public uint Length { get; set; }
+    public long Length { get; set; }
 
     /// <summary>
     ///     The width of this complete object in pixels, including follow-on portions.
     /// </summary>
-    public ushort Width { get; set; }
+    public int Width { get; set; }
 
     /// <summary>
     ///     The height of this complete object in pixels, including follow-on portions.
     /// </summary>
-    public ushort Height { get; set; }
+    public int Height { get; set; }
 
     /// <summary>
     ///     Initializes a new instance with default values including an empty data buffer.
@@ -72,7 +72,7 @@ public class InitialObjectDefinitionSegment : ObjectDefinitionSegment
     ///     The RLE-compressed data for this portion of the completed object.
     /// </param>
     public InitialObjectDefinitionSegment(PgsTimeStamp pts, PgsTimeStamp dts
-        , VersionedId<ushort> versionedId, ushort width, ushort height, uint length
+        , VersionedId<int> versionedId, int width, int height, long length
         , byte[] data) : base(pts, dts, versionedId, data)
     {
         Length = length;
